♻️ Refactor protocols status logic
