+++ pages/_app.tsx
@@ -1,7 +1,7 @@
 import type { AppProps } from 'next/app';
 import '../styles/globals.css';
 
-import Providers from 'components/Providers';
+import Providers from 'components/Providers/Providers';
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
